@@ -1,6 +1,6 @@
 class ListadoController < ApplicationController
 
-before_filter :login_requerido
+before_action :login_requerido
 
   def asignacion_lectivo_impresa
     @asignacions = Asignaciondef.all
