@@ -2,7 +2,7 @@ class SolicitudrecursousuariosController < ApplicationController
  # GET /solicitudrecursos
   # GET /solicitudrecursos.xml
 
-  before_filter :login_requerido,:usuario?
+  before_action :login_requerido,:usuario?
 
   def index
     @solicitudrecursos= Solicitudrecurso.where("usuario_id = ?",session[:user_id]).to_a
