@@ -6,7 +6,7 @@ Gestorlcc::Application.configure do
   # your test database is "scratch space" for the test suite and is wiped
   # and recreated between test runs.  Don't rely on the data there!
   config.cache_classes = true
-
+config.active_support.test_order = :sorted
   # Log error messages when you accidentally call methods on nil.
   #config.whiny_nils = true
 
